+++ app/Http/Controllers/TagsController.php
@@ -105,6 +105,11 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        //~Tags cannot be deleted because its associated to some posts
+        if($tag->count() > 0) {
+            session()->flash('error', 'Tag cannot be deleted, because it is associated to some posts.');
+            return redirect()->back();
+        }
         //~Delete data to database
         $tag->delete();
